Clarifies handling presence of both 1.0 and 2.0 strings
This is a requested clarification of how CMPs should handle the simultaneous presence of v1.x strings and v2.x strings during a transition period where both could erroneously be present at the same time. The goal should be to upgrade the string to the latest version which most accurately reflects user choices.